+++ src/main/java/com/hryshchenko/cinema/model/entity/Ticket.java
@@ -2,32 +2,32 @@ package com.hryshchenko.cinema.model.entity;
 
 public class Ticket extends Entity {
     private static final long serialVersionUID = 1L;
-    private int screeningId;
-    private int userId;
+    private long screeningId;
+    private long userId;
     private int ticketCount;
 
     public Ticket() {}
 
-    public Ticket(long id, int screeningId, int userId, int ticketCount) {
+    public Ticket(long id, long screeningId, long userId, int ticketCount) {
         super(id);
         this.screeningId = screeningId;
         this.userId = userId;
         this.ticketCount = ticketCount;
     }
 
-    public int getScreeningId() {
+    public long getScreeningId() {
         return screeningId;
     }
 
-    public void setScreeningId(int screeningId) {
+    public void setScreeningId(long screeningId) {
         this.screeningId = screeningId;
     }
 
-    public int getUserId() {
+    public long getUserId() {
         return userId;
     }
 
-    public void setUserId(int userId) {
+    public void setUserId(long userId) {
         this.userId = userId;
     }
